+++ .circleci/example-green-project/src/example_project/core.clj
@@ -1,0 +1,1 @@
+(ns example-project.core)