+++ .circleci/example-red-project/src/example_project/core.clj
@@ -1,0 +1,3 @@
+(ns example-project.core)
+
+clojure.set/difference